+++ src/Guilded.Base/server/MemberSummary.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using Guilded.Base.Users;
 using Newtonsoft.Json;
 
@@ -35,11 +37,20 @@ public class MemberSummary<T> : BaseModel where T : UserSummary
     /// <seealso cref="Id" />
     public IList<uint> RoleIds { get; }
 
+    /// <summary>
+    /// Gets the identifier of the <see cref="Server">server</see> where the <see cref="Member">member</see> is.
+    /// </summary>
+    /// <value><see cref="Server.Id">Server ID</see></value>
+    public HashId ServerId { get; }
+
     /// <inheritdoc cref="UserSummary.Id" />
     public HashId Id => User.Id;
 
     /// <inheritdoc cref="UserSummary.Name" />
     public string Name => User.Name;
+
+    /// <inheritdoc cref="UserSummary.Avatar" />
+    public Uri? Avatar => User.Avatar;
 
     /// <inheritdoc cref="UserSummary.Type" />
     public UserType Type => User.Type;
@@ -54,14 +65,59 @@ public class MemberSummary<T> : BaseModel where T : UserSummary
     /// </summary>
     /// <param name="user"><see cref="Users.User">The user</see> who is present in <see cref="Server">the server</see></param>
     /// <param name="roleIds">The list of roles user holds</param>
+    /// <param name="serverId">The identifier of the <see cref="Server">server</see> where the <see cref="Member">member</see> is</param>
     /// <returns>New <see cref="MemberSummary{T}" /> JSON instance</returns>
     /// <seealso cref="MemberSummary{T}" />
     public MemberSummary(
         T user,
 
-        IList<uint> roleIds
+        IList<uint> roleIds,
+
+        HashId serverId
     ) =>
-        (User, RoleIds) = (user, roleIds);
+        (User, RoleIds, ServerId) = (user, roleIds, serverId);
+    #endregion
+
+    #region Methods
+    /// <inheritdoc cref="BaseGuildedClient.GetSocialLinkAsync(HashId, HashId, SocialLinkType)" />
+    public Task<SocialLink> GetSocialLinkAsync(SocialLinkType linkType) =>
+        User.GetSocialLinkAsync(ServerId, linkType);
+
+    /// <inheritdoc cref="BaseGuildedClient.UpdateNicknameAsync(HashId, HashId, string)" />
+    public Task<string> UpdateNicknameAsync(string nickname) =>
+        User.UpdateNicknameAsync(ServerId, nickname);
+
+    /// <inheritdoc cref="BaseGuildedClient.DeleteNicknameAsync(HashId, HashId)" />
+    public Task DeleteNicknameAsync() =>
+        User.DeleteNicknameAsync(ServerId);
+
+    /// <inheritdoc cref="BaseGuildedClient.AddRoleAsync(HashId, HashId, uint)" />
+    public Task AddRoleAsync(uint role) =>
+        User.AddRoleAsync(ServerId, role);
+
+    /// <inheritdoc cref="BaseGuildedClient.RemoveRoleAsync(HashId, HashId, uint)" />
+    public Task RemoveRoleAsync(uint role) =>
+        User.RemoveRoleAsync(ServerId, role);
+
+    /// <inheritdoc cref="BaseGuildedClient.AddXpAsync(HashId, HashId, short)" />
+    public Task<long> AddXpAsync(short amount) =>
+        User.AddXpAsync(ServerId, amount);
+
+    /// <inheritdoc cref="BaseGuildedClient.RemoveMemberAsync(HashId, HashId)" />
+    public Task RemoveMemberAsync() =>
+        User.RemoveMemberAsync(ServerId);
+
+    /// <inheritdoc cref="BaseGuildedClient.AddMemberBanAsync(HashId, HashId, string?)" />
+    public Task AddMemberBanAsync(string? reason = null) =>
+        User.AddMemberBanAsync(ServerId, reason);
+
+    /// <inheritdoc cref="BaseGuildedClient.RemoveMemberBanAsync(HashId, HashId)" />
+    public Task RemoveMemberBanAsync() =>
+        User.RemoveMemberBanAsync(ServerId);
+
+    /// <inheritdoc cref="BaseGuildedClient.GetBanAsync(HashId, HashId)" />
+    public Task GetBanAsync() =>
+        User.GetBanAsync(ServerId);
     #endregion
 }
 
@@ -78,8 +134,9 @@ public class MemberSummary : MemberSummary<UserSummary>
     /// <summary>
     /// Initializes a new instance of <see cref="MemberSummary{T}" />.
     /// </summary>
-    /// <param name="user"><see cref="Users.User">The user</see> who is present in <see cref="Server">the server</see></param>
+    /// <param name="user"><see cref="User">The user</see> who is present in <see cref="Server">the server</see></param>
     /// <param name="roleIds">The list of roles user holds</param>
+    /// <param name="serverId">The identifier of the <see cref="Server">server</see> where the <see cref="Member">member</see> is</param>
     /// <returns>New <see cref="MemberSummary{T}" /> JSON instance</returns>
     /// <seealso cref="MemberSummary{T}" />
     [JsonConstructor]
@@ -88,7 +145,10 @@ public class MemberSummary : MemberSummary<UserSummary>
         UserSummary user,
 
         [JsonProperty(Required = Required.Always)]
-        IList<uint> roleIds
-    ) : base(user, roleIds) { }
+        IList<uint> roleIds,
+
+        [JsonProperty(Required = Required.Always)]
+        HashId serverId
+    ) : base(user, roleIds, serverId) { }
     #endregion
 }